install jest; rename some code named spec.ts so jest doesn't pick it up.
