+++ src/packages/frontend/frame-editors/whiteboard-editor/elements/desc.ts
