+++ src/packages/frontend/frame-editors/whiteboard-editor/tools/desc.ts
