+++ src/packages/frontend/frame-editors/whiteboard-editor/tools/tool-panel.tsx
@@ -27,9 +27,9 @@ import IconSelect from "@cocalc/frontend/components/icon-select";
 import { getCountdownMoment } from "@cocalc/frontend/editors/stopwatch/stopwatch";
 import { AspectRatio } from "./frame";
 import { ResetButton, SELECTED } from "./common";
-import { Tool, TOOLS } from "./spec";
+import { Tool, TOOLS } from "./desc";
 export type { Tool };
-import { ELEMENTS } from "../elements/spec";
+import { ELEMENTS } from "../elements/desc";
 import { Element } from "../types";
 import { redux } from "@cocalc/frontend/app-framework";
 import { set_account_table } from "@cocalc/frontend/account/util";
