+++ src/packages/frontend/package.json
@@ -182,12 +182,14 @@
     "@types/react-redux": "^7.1.9",
     "cjsx-register": "^1.1.0",
     "coffeescript": "^2.5.1",
+    "jest": "^29.3.1",
     "mocha": "^10.0.0",
     "typescript": "^4.7.4"
   },
   "scripts": {
     "build": "rm -rf dist && NODE_OPTIONS=--max-old-space-size=8192 tsc && coffee  -m -c -o dist/ .",
-    "tsc": "npx tsc  --watch  --pretty --preserveWatchOutput "
+    "tsc": "npx tsc  --watch  --pretty --preserveWatchOutput ",
+    "test": "npx jest dist/"
   },
   "homepage": "https://github.com/sagemathinc/cocalc",
   "repository": {
